Refactor; move `ExtendedXXX` to separate file
